Safari-611.1.10/LayoutTests/fast/dom/timer-throttling-hidden-page-non-nested-expected.txt

Tests that DOM timers on hidden pages that have not reached max nesting level are not throttled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.isTimerThrottled(timerHandle) is false
PASS internals.isTimerThrottled(timerHandle) is false
PASS internals.isTimerThrottled(timerHandle) is false
PASS internals.isTimerThrottled(timerHandle) is false
PASS internals.isTimerThrottled(timerHandle) is false
PASS successfullyParsed is true

TEST COMPLETE
